ARDd: Fix bug with closing instance if there's no map image
git-svn-id: https://itsd.ict.sait.ca:9880/capstone07/trunk@320 0c6396e7-dfa7-ad41-aac3-ea9b0ac9ed18

+++ ardd/ardd/src/main/java/opticnav/ardd/instance/Instance.java
@@ -25,7 +25,9 @@ public class Instance implements AutoCloseable {
     
     @Override
     public void close() throws IOException {
-        this.info.mapImage.delete();
+        if (this.info.mapImage != null) {
+            this.info.mapImage.delete();
+        }
     }
     
     public InstanceInfo getInfo() {
